Rename all performance/rating graph related classes, files and functions
