+++ Quaver.Shared/Screens/Results/UI/Tabs/Overview/Graphs/ResultsOverviewGraphContainer.cs
@@ -16,7 +16,7 @@ using Quaver.Shared.Screens.Results.UI.Tabs.Overview.Graphs.Accuracy;
 using Quaver.Shared.Screens.Results.UI.Tabs.Overview.Graphs.Deviance;
 using Quaver.Shared.Screens.Results.UI.Tabs.Overview.Graphs.Footer;
 using Quaver.Shared.Screens.Results.UI.Tabs.Overview.Graphs.Health;
-using Quaver.Shared.Screens.Results.UI.Tabs.Overview.Graphs.Performance;
+using Quaver.Shared.Screens.Results.UI.Tabs.Overview.Graphs.Rating;
 using Quaver.Shared.Screens.Selection.UI.FilterPanel.MapInformation.Metadata;
 using Wobble;
 using Wobble.Assets;
@@ -306,7 +306,7 @@ namespace Quaver.Shared.Screens.Results.UI.Tabs.Overview.Graphs
                 CreateDevianceGraph();
                 CreateAccuracyGraph();
                 CreateHealthGraph();
-                CreatePerformanceGraph();
+                CreateRatingGraph();
                 ToggleGraphVisibility();
                 return;
             }
@@ -354,7 +354,7 @@ namespace Quaver.Shared.Screens.Results.UI.Tabs.Overview.Graphs
         /// <summary>
         ///     Creates/Enables the performance graph
         /// </summary>
-        private void CreatePerformanceGraph() => Graphs[ResultGraphs.Rating] = new CachedPerformanceGraph(Map, Processor, GraphSize)
+        private void CreateRatingGraph() => Graphs[ResultGraphs.Rating] = new CachedRatingGraph(Map, Processor, GraphSize)
         {
             Parent = GraphContainer,
             Alignment = Alignment.MidCenter,
